requirements.txt
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
botocore==1.37.3
certifi==2025.4.26
click==8.1.8
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
docstring_parser==0.16
executing==2.2.0
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ijson==3.4.0
inspect_ai==0.3.103
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jmespath==1.0.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
linkify-it-py==2.0.3
markdown-it-py==3.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mmh3==5.1.0
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.3.0
openai==1.86.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
ply==3.11
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
rich==13.9.4
rpds-py==0.25.1
s3fs==2025.3.0
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
tenacity==9.1.2
textual==2.1.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.4.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
